#51: Fix celery tasks discovery
